# functional node_density implementation