functional implementation of nodedensity.
I kept the function logic pretty much the same, just used the new graph api and np.in1d, instead of np.isin.

Timings:
old:
CPU times: user 14.9 s, sys: 69.3 ms, total: 15 s
Wall time: 14.9 s

new:
CPU times: user 6.19 s, sys: 0 ns, total: 6.19 s
Wall time: 6.19 s

Potentially, the function can be sped up further if we change the function API and pass the networkx.MultiGraph object.
Then, we can do:

sum(l for (_,_,l) in nx.subgraph(neighbours).edges.data('mm_len'))

where neighbours comes from the new libpysal.Graph.