@jGaboardi 2024-05-16
> Potentially, the function can be sped up further if we change the function API and pass the networkx.MultiGraph object.
Then, we can do:
>
> `sum(l for (,,l) in nx.subgraph(neighbours).edges.data('mm_len'))`
>
> where neighbours comes from the new libpysal.Graph.

Was there any more thought to this?