@martinfleis 2024-05-16
> Was there any more thought to this?

Yes. It was faster but the API got super weird so we scraped it.